% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{db_close}
\alias{db_close}
\title{Close database connection}
\usage{
db_close(conn)
}
\arguments{
\item{conn}{RODBC connection object as returned by db_connect()}
}
\description{
Close database connection
}
