% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcp.r
\name{db_connect}
\alias{db_connect}
\title{Open database connection. Connection parameters are read from options. Database can be overriden by passing the \code{db} parameter.}
\usage{
db_connect(db = getOption("sqlservr.db_database"))
}
\arguments{
\item{db}{Name of database to connect to}
}
\value{
RODBC connection object
}
\description{
Open database connection. Connection parameters are read from options. Database can be overriden by passing the \code{db} parameter.
}
