% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlservr.r
\name{db_query}
\alias{db_query}
\title{Execute SQL query from file}
\usage{
db_query(file, query = 1, parameters = NULL, conn = NULL)
}
\arguments{
\item{file}{filename of file containing the SQL query}

\item{query}{character string referencing the SQL query in the file (in case there is more than one)}

\item{parameters}{list of parameters to substitute in the SQL query}

\item{conn}{RODBC connection object as returned by db_connect(). Optional, but allows to re-use an already open connection.}
}
\value{
SQL query as character string
}
\description{
Execute SQL query from file
}
