% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcp.r
\name{get_sql}
\alias{get_sql}
\title{Get SQL query from file}
\usage{
get_sql(file, query = 1, parameters = NULL)
}
\arguments{
\item{file}{filename of file containing the SQL query}

\item{query}{character string referencing the SQL query in the file (in case there is more than one)}

\item{parameters}{list of parameters to substitute in the SQL query}
}
\value{
SQL query as character string
}
\description{
Get SQL query from file
}
