% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_sql_queries}
\alias{get_sql_queries}
\title{Retrieves reference and name of available queries from a file}
\usage{
get_sql_queries(file)
}
\arguments{
\item{file}{filename of file containing SQL queries}
}
\description{
Retrieves reference and name of available queries from a file
}
