% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcp.r
\name{db_bcp}
\alias{db_bcp}
\title{Write data.frame to database using bcp}
\usage{
db_bcp(data, table, conn = NULL, truncate = FALSE, tmpSuffix = "tmp",
  tmpDir = ".")
}
\arguments{
\item{data}{data.frame to be written to database. Column names must match with those in the database table.}

\item{table}{character string naming the table to write to. Can include database and schema name (database_name.schema_name.table_name).}

\item{conn}{RODBC connection object as returned by db_connect(). Optional, but allows to re-use an already open connection to check columns and truncate table (if \code{truncate} is \code{TRUE}).}

\item{truncate}{logical. If \code{TRUE}, the table will be truncated prior to writing data.}

\item{tmpSuffix}{character string}

\item{tmpDir}{character string}
}
\description{
Write data.frame to database using bcp
}
